create date to locale string util
